+++ .idea/artifacts/api.xml
@@ -1,6 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="api">
     <output-path>$PROJECT_DIR$/api/build/libs</output-path>
-    <root id="archive" name="api.jar" />
+    <root id="archive" name="api.jar">
+      <element id="module-output" name="Valiutchik.api.androidMain" />
+    </root>
   </artifact>
 </component>
