+++ api/build.gradle.kts
@@ -88,6 +88,7 @@ kotlin {
                 implementation(libs.ktor.auth)
                 implementation(libs.ktor.content)
                 implementation(libs.ktor.encoding)
+                implementation(libs.ktor.logging)
                 implementation(libs.ktor.serialization)
                 implementation(libs.ktor.serialization.xml)
                 implementation(libs.kotlinx.serialization)
