+++ api/src/commonMain/kotlin/fobo66/valiutchik/api/di/NetworkModule.kt
@@ -16,27 +16,38 @@
 
 package fobo66.valiutchik.api.di
 
+import io.github.aakira.napier.Napier
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.cio.CIO
-import io.ktor.client.plugins.cache.HttpCache
-import io.ktor.client.plugins.cache.storage.FileStorage
 import io.ktor.client.plugins.compression.ContentEncoding
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.http.HttpHeaders
 import io.ktor.serialization.kotlinx.json.json
 import io.ktor.serialization.kotlinx.xml.xml
 import kotlinx.serialization.json.Json
-import org.koin.android.ext.koin.androidContext
+import nl.adaptivity.xmlutil.ExperimentalXmlUtilApi
 import org.koin.dsl.module
 
+@OptIn(ExperimentalXmlUtilApi::class)
 val networkModule =
     module {
+        single<Logger> {
+            object : Logger {
+                override fun log(message: String) {
+                    Napier.d(message, tag = "Ktor")
+                }
+            }
+        }
         single<Json> {
             Json {
                 isLenient = true
                 ignoreUnknownKeys = true
             }
         }
-        single {
+        single<HttpClient> {
             HttpClient(CIO) {
                 install(ContentNegotiation) {
                     json(get())
@@ -46,8 +57,10 @@ val networkModule =
                     gzip()
                     deflate()
                 }
-                install(HttpCache) {
-                    publicStorage(FileStorage(androidContext().cacheDir))
+                install(Logging) {
+                    logger = get()
+                    level = LogLevel.ALL
+                    sanitizeHeader { header -> header == HttpHeaders.Authorization }
                 }
 
                 expectSuccess = true
